modularize main function into graphics test
