+++ src/final/graphicstest.h
@@ -1,0 +1,24 @@
+#ifndef graphicstest_h
+#define graphicstest_h
+
+#include <stdint.h>
+
+void graphics_test_all();
+void run_all_benchmarks();
+
+// individual benchmarks
+unsigned long testFillScreen();
+unsigned long testText();
+unsigned long testLines(uint16_t color);
+unsigned long testFastLines(uint16_t color1, uint16_t color2);
+unsigned long testRects(uint16_t color);
+unsigned long testFilledRects(uint16_t color1, uint16_t color2);
+unsigned long testFilledCircles(uint8_t radius, uint16_t color);
+unsigned long testCircles(uint8_t radius, uint16_t color);
+unsigned long testTriangles();
+unsigned long testFilledTriangles();
+unsigned long testRoundRects();
+unsigned long testFilledRoundRects();
+void test_rotation(void);
+
+#endif
